funcion para google cloud datasets comunes/requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
openpyxl
numpy
pyarrow